+++ src/main/java/net/fabricmc/loom/LoomGradlePlugin.java
@@ -30,7 +30,7 @@ import java.util.Locale;
 import org.gradle.api.Project;
 import org.gradle.api.tasks.TaskContainer;
 
-import net.fabricmc.loom.decompilers.LoomDecompiler;
+import net.fabricmc.loom.api.decompilers.LoomDecompiler;
 import net.fabricmc.loom.decompilers.fernflower.FabricFernFlowerDecompiler;
 import net.fabricmc.loom.providers.MappingsProvider;
 import net.fabricmc.loom.task.CleanEclipseRunsTask;
